Almost first running version

+++ src/main/logic/IAnimalStatusChangeObserver.java
@@ -1,6 +1,6 @@
 package logic;
 
 public interface IAnimalStatusChangeObserver {
-    void energyChanged(Animal prevState,int newEnergy);
-    void positionChanged(Animal prevState,Vector2d newPosition, MapDirection newOrientation,int newEnergy);
+    void energyChanged(Animal prevState,double newEnergy);
+    void positionChanged(Animal prevState,Vector2d newPosition, MapDirection newOrientation,double newEnergy);
 }
